test(provider-node): move coordinator mock tests to integration tests for CI coverage
Move 29 coordinator mock tests from #[cfg(test)] blocks in source files
to integration test files under provider-node/tests/. CI measures
provider-node coverage from integration tests only, so these tests were
invisible to the coverage gate despite exercising production code paths.

- Create 4 integration test files (agreement, challenge, checkpoint, replica)
- Remove #[cfg(test)] blocks from the 4 coordinator source files
- Make tested methods pub so integration tests can call them
- Use newtype SharedMock wrappers to satisfy orphan rules
- Add new test targets to CI coverage discovery loop in check.yml
